compliance_tool/requirements.txt
matplotlib>=3.8.0
pillow>=10.2.0
python-docx>=1.1.0
